# Please enter a commit message to explain why this merge is necessary,
# especially if it merges an updated upstream into a topic branch.
#
# Lines starting with '#' will be ignored, and an empty message aborts
# the commit.

+++ Project/src/Payment.java
@@ -49,7 +49,7 @@ public class Payment
 			if(paymentMethod.equalsIgnoreCase("card"))
 			{
 				System.out.print("Please input the card number: ");
-				int cardNumber = inputReader.nextInt();
+				long cardNumber = inputReader.nextInt();
 				System.out.println("Please enter your pin number: ");
 				int pinNumber = inputReader.nextInt();
 				paymentCorrect = true;
